Funcion de buscar,cambios en las vistas y funcion de eliminar arreglada

+++ src/routes/evento.js
@@ -65,9 +65,19 @@ router.get('/', isLoggedIn, roles,async (req, res) => {
 
 router.get('/delete/:id', async (req, res) => {
     const { id } = req.params;
-    await pool.query('DELETE FROM eventos WHERE ID = ?', [id]);
-    req.flash('success', 'Evento eliminado');
-    res.redirect('/evento');
+    const evento = await pool.query('SELECT * FROM bitacoras WHERE eventoId = ? ',[id]);
+    if (evento[0]==null){
+        const { id } = req.params;
+        await pool.query('DELETE FROM eventos WHERE ID = ?', [id]);
+        req.flash('success', 'Evento eliminado');
+        res.redirect('/evento');
+    
+}else{
+    req.flash('message', 'ERROR, este campo no puede ser eliminado');
+res.redirect('/evento');
+
+}
+  
 });
 
 router.get('/edit/:id', isLoggedIn,roles,async (req, res) => {
@@ -102,5 +112,30 @@ router.post('/edit/:id', async (req, res) => {
     req.flash('success', 'Evento actualizado');
     res.redirect('/evento');
 });
+router.post('/buscar', isLoggedIn,roles, async (req, res) => {
+    var loginAdmin = false;
+    var loginGeneral = false;
+
+    rol = req.user.rol;
+    if(rol == "Admin") {
+        loginAdmin = true;
+        }
+        if(rol == "General") {
+            loginGeneral = true;
+            }
+
+     const { evento } = req.body;
+    const eventos = await pool.query('SELECT * FROM eventos WHERE evento = ?', [evento]);
+    
+    res.render('evento/buscar', { 
+        
+        layout: "dashboard",
+        loginAdmin,
+        loginGeneral,
+        eventos 
+    });
+});
+
+
 
 module.exports = router;
