+++ src/routes/incidente.js
@@ -63,11 +63,21 @@ router.get('/', isLoggedIn, roles,async (req, res) => {
     });
 });
 
-router.get('/delete/:id', isLoggedIn,roles,async (req, res) => {
+router.get('/delete/:id', async (req, res) => {
     const { id } = req.params;
-    await pool.query('DELETE FROM incidentes WHERE ID = ?', [id]);
-    req.flash('success', 'incidente eliminado');
-    res.redirect('/incidente');
+    const incidente = await pool.query('SELECT * FROM plataformas WHERE incidenteId = ? ',[id]);
+    if (incidente[0]==null){
+        const { id } = req.params;
+        await pool.query('DELETE FROM incidentes WHERE ID = ?', [id]);
+        req.flash('success', 'Incidente eliminado');
+        res.redirect('/incidente');
+    
+}else{
+    req.flash('message', 'ERROR, este campo no puede ser eliminado');
+res.redirect('/incidente');
+
+}
+  
 });
 
 router.get('/edit/:id', isLoggedIn,roles,async (req, res) => {
@@ -103,5 +113,31 @@ router.post('/edit/:id', async (req, res) => {
     req.flash('success', 'incidente actualizado');
     res.redirect('/incidente');
 });
+
+router.post('/buscar', isLoggedIn,roles, async (req, res) => {
+    var loginAdmin = false;
+    var loginGeneral = false;
+
+    rol = req.user.rol;
+    if(rol == "Admin") {
+        loginAdmin = true;
+        }
+        if(rol == "General") {
+            loginGeneral = true;
+            }
+
+     const { incidente } = req.body;
+    const incidentes = await pool.query('SELECT * FROM incidentes WHERE incidente = ?', [incidente]);
+    
+    res.render('incidente/buscar', { 
+        
+        layout: "dashboard",
+        loginAdmin,
+        loginGeneral,
+        incidentes 
+    });
+});
+
+
 
 module.exports = router;
